Add support for short Git repository schema
The Git repository schema can now understand repositories of the form:

```json
{
    "remotes": {
        "origin": "git@example.com:example.git"
    }
}
```

i.e., repository remotes that consist only of a URL.

Multiple remotes of this sort can also be specified, as long as one can
be determined to be the default (i.e., one is named "origin").

Under the hood, we have a new field called an `EitherField` that maps
input/output types to marshmallow fields for (de)serialization. We use
this to map a repository to a data structure that keeps track of whether
or not it was originally an implicit (i.e., only a URL) or explicit
(i.e., a dictionary with a URL and other fields) repository. This is
mapped back to the underlying data structure (`str` or `dict`) during
serialization.

With these changes, we now specify the field that errored when
validating a `GitRepository`, so if the repository is missing a default
remote that will show up assigned to the `remotes` field instead of on
the fallback `_schema` field.